rename commit_edit to version
